Moves GenerateResponse to its own file.
