+++ js/ai/src/generateAction.ts
@@ -25,11 +25,9 @@ import { toJsonSchema } from '@genkit-ai/core/schema';
 import { runInNewSpan, SPAN_TYPE_ATTR } from '@genkit-ai/core/tracing';
 import * as clc from 'colorette';
 import { DocumentDataSchema } from './document.js';
-import {
-  GenerateResponse,
-  GenerateResponseChunk,
-  tagAsPreamble,
-} from './generate.js';
+import { tagAsPreamble } from './generate.js';
+import { GenerateResponseChunk } from './generate/chunk.js';
+import { GenerateResponse } from './generate/response.js';
 import {
   GenerateRequest,
   GenerateRequestSchema,
